+++ Imperium/src/Interface/ImperiumUI/Windows/MoonControl/MoonControlWindow.cs
@@ -5,6 +5,9 @@ using Imperium.Interface.Common;
 using Imperium.Interface.ImperiumUI.Windows.MoonControl.Widgets;
 using Imperium.MonoBehaviours.ImpUI.Common;
 using Imperium.Types;
+using Imperium.Util;
+using TMPro;
+using UnityEngine;
 
 #endregion
 
@@ -12,10 +15,16 @@ namespace Imperium.Interface.ImperiumUI.Windows.MoonControl;
 
 internal class MoonControlWindow : ImperiumWindow
 {
+    private TMP_InputField levelSeedInput;
+    private TMP_Text levelSeedTitle;
+    private TMP_Text levelSeedText;
+
     protected override void InitWindow()
     {
         InitSpawnPropertyFields();
         InitMapObstacleButtons();
+        InitEntitySpawning();
+        InitGeneration();
 
         RegisterWidget<WeatherForecaster>(transform, "Right/Weather");
         RegisterWidget<TimeManipulation>(transform, "Right/Time");
@@ -28,6 +37,51 @@ internal class MoonControlWindow : ImperiumWindow
             transform,
             new StyleOverride("Border", Variant.DARKER),
             new StyleOverride("Right", Variant.DARKER)
+        );
+    }
+
+    private void InitEntitySpawning()
+    {
+        ImpToggle.Bind(
+            "Left/PauseIndoorSpawning", transform,
+            Imperium.MoonManager.IndoorSpawningPaused,
+            theme
+        );
+
+        ImpToggle.Bind(
+            "Left/PauseOutdoorSpawning", transform,
+            Imperium.MoonManager.OutdoorSpawningPaused,
+            theme
+        );
+
+        ImpToggle.Bind(
+            "Left/PauseDaytimeSpawning", transform,
+            Imperium.MoonManager.DaytimeSpawningPaused,
+            theme
+        );
+    }
+
+    private void InitGeneration()
+    {
+        levelSeedTitle = transform.Find("Right/Generation/Seed/Title").GetComponent<TMP_Text>();
+        levelSeedText = transform.Find("Right/Generation/Seed/Input/Text Area/Text").GetComponent<TMP_Text>();
+
+        ImpButton.Bind(
+            "Right/Generation/Seed/Reset",
+            transform,
+            () => Imperium.GameManager.CustomSeed.Reset(),
+            theme: theme,
+            interactableInvert: true,
+            interactableBindings: Imperium.IsSceneLoaded
+        );
+
+        levelSeedInput = ImpInput.Bind(
+            "Right/Generation/Seed/Input",
+            transform,
+            Imperium.GameManager.CustomSeed,
+            theme: theme,
+            interactableInvert: true,
+            interactableBindings: Imperium.IsSceneLoaded
         );
     }
 
@@ -182,5 +236,22 @@ internal class MoonControlWindow : ImperiumWindow
             interactableBindings: Imperium.IsSceneLoaded,
             theme: theme
         );
+    }
+
+    protected override void OnOpen()
+    {
+        levelSeedInput.text = Imperium.IsSceneLoaded.Value
+            ? Imperium.StartOfRound.randomMapSeed.ToString()
+            : Imperium.GameManager.CustomSeed.Value != -1
+                ? Imperium.GameManager.CustomSeed.Value.ToString()
+                : "";
+
+        Imperium.GameManager.ProfitQuota.Refresh();
+        Imperium.GameManager.GroupCredits.Refresh();
+        Imperium.GameManager.QuotaDeadline.Refresh();
+
+        levelSeedInput.interactable = !Imperium.IsSceneLoaded.Value;
+        ImpUtils.Interface.ToggleTextActive(levelSeedTitle, !Imperium.IsSceneLoaded.Value);
+        ImpUtils.Interface.ToggleTextActive(levelSeedText, !Imperium.IsSceneLoaded.Value);
     }
 }