+++ Imperium/src/Visualizers/SpawnIndicators.cs
@@ -20,16 +20,16 @@ internal class SpawnIndicators(
     {
         ClearObjects();
 
-        for (var i = state.currentCycle; i < state.outdoorCycles.Length; i++)
+        for (var i = state.CurrentCycle; i < state.OutdoorCycles.Length; i++)
         {
-            foreach (var spawnReport in state.outdoorCycles[i])
+            foreach (var spawnReport in state.OutdoorCycles[i])
             {
                 var indicatorObject = Object.Instantiate(ImpAssets.SpawnIndicator);
                 var indicator = indicatorObject.AddComponent<SpawnIndicator>();
-                indicator.transform.position = spawnReport.position;
+                indicator.transform.position = spawnReport.Position;
                 indicator.Init(
-                    Imperium.ObjectManager.GetDisplayName(spawnReport.entity.enemyName),
-                    spawnReport.spawnTime
+                    Imperium.ObjectManager.GetDisplayName(spawnReport.Entity.enemyName),
+                    spawnReport.SpawnTime
                 );
 
                 visualizerObjects[indicatorObject.GetInstanceID()] = indicator;